+++ index.ts
@@ -1,6 +1,6 @@
 import * as camelcasify from './parsers/camelcasify/camelcasify.parser';
 import * as pick from './parsers/pick/pick.parser';
 import * as snakecasify from './parsers/snakecasify/snakecasify.parser';
-import * as toCss from './parsers/to-css/to-css.parser';
+import * as toCss from './parsers/to-css-custom-properties/to-css-custom-properties.parser';
 
 export { camelcasify, pick, snakecasify, toCss };
