+++ parsers/to-css-custom-properties/tokens/depth.ts
