+++ parsers/to-css-custom-properties/tokens/duration.ts
