+++ parsers/to-css-custom-properties/tokens/textStyle.ts
