+++ parsers/to-css-custom-properties/tokens/vector.ts
